[extension] separate the addTournesolRecommendations content script into two new content script displayHomeRecommendations and displaySearchRecommendation
