+++ browser-extension/src/manifest.json
@@ -34,7 +34,7 @@
   "content_scripts": [
     {
       "matches": ["https://*.youtube.com/*"],
-      "js": ["addTournesolRecommendations.js"],
+      "js": ["displayHomeRecommendations.js","displaySearchRecommendations.js"],
       "css": ["addTournesolRecommendations.css"],
       "run_at": "document_start",
       "all_frames": true
