+++ browser-extension/src/models/banner/Banner.js
@@ -3,10 +3,10 @@ import { defaultBannerOptions } from './BannerOptions.js';
 
 /**
  * A banner displaying a message and an action button.
- * 
+ *
  * At the moment this class is specific to the extension research study, but
  * could be refactored to be generic.
- * 
+ *
  * - the texts could be parameters of the constructor
  * - the action button could be optional
  * - etc.
