+++ browser-extension/src/models/banner/BannerOptions.js
